frees in mlx refac
