+++ tests/integration_tests/testfiles/valid_nocy.rt
@@ -1,4 +1,4 @@
 A 0.2 255,255,255
 C 0,0,1 0,0,1 170
 L -40,0,30 0.7 255,255,255
-sp 0,0,5 10 0,255,0
+sp 0,0,5 10 124,213,77
